refactor: move types to avoid importing worker file
