+++ packages/@sanity/cli/src/actions/schema/validateSchema.worker.ts
@@ -14,6 +14,7 @@ import {
 } from '@sanity/types'
 
 import {getWorkspace} from '../../util/getWorkspace.js'
+import {type SerializedSchemaDebug, type SerializedTypeDebug} from './types.js'
 
 /** @internal */
 export interface ValidateSchemaWorkerData {
@@ -29,30 +30,6 @@ export interface ValidateSchemaWorkerResult {
   validation: SchemaValidationProblemGroup[]
 
   serializedDebug?: SerializedSchemaDebug
-}
-
-/**
- * Contains debug information about the serialized schema.
- *
- * @internal
- **/
-export type SerializedSchemaDebug = {
-  hoisted: Record<string, SerializedTypeDebug>
-  parent?: SerializedSchemaDebug
-  size: number
-  types: Record<string, SerializedTypeDebug>
-}
-
-/**
- * Contains debug information about a serialized type.
- *
- * @internal
- **/
-export type SerializedTypeDebug = {
-  extends: string
-  fields?: Record<string, SerializedTypeDebug>
-  of?: Record<string, SerializedTypeDebug>
-  size: number
 }
 
 const {
